Move server to root directory
